pyproject.toml
[project]
name = "mcp-coupang-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "mcp[cli]>=1.19.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
